Consolidated tests into shared base.
